[ansible] PoC Direct image download from Glance
as Ansible deploy driver is communicating with the node over encrypted
transport (SSH) it is not that great security risk to pass in a token
that can be used to download the image from the Glance API directly.

To be more secure, a new ironic-specific keystone domain is introduced.
It must be set up during ironic deployment together with admin user
inside this domain, and details of those domain and admin user set in
ironic.conf.

When deploying, a new keystone user is created inside the ironic domain,
and a trust is created between ironic service user and this temporary
user. This trust is then used to create a token to pass to deploy
ramdisk.

On tearing down the deployment, the trust is deleted (effectively
invalidating the issued token), and the temporary user is also removed.

Change-Id: I280b5063694448bfb26e43f04022317a5dae825b